requirements.txt
asgiref==3.2.10
Django==3.0.8
django-modeltranslation==0.15.1
django-taggit==1.3.0
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
